Tweak UI and add in event dimensions
This commit modifies the UI. The UI lists the API and also the list of events and its floor plan in SVG.
The app renders the event data using EJS. Dockerfile is updated to copy the views folder containing the ejs file.
This commit also adds in the dimensions for an event in the model and mock data.
Modified README for better flow.

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>