Update Mongoose and Update Compose steps
This commit updates the version of mongoose to the current latest (5.2).

This commit also updates the use of Compose for MongoDB. The service is
now using Let's Encrypt certificates. Previously, a provided certificate
was needed to be mounted for connection. The MongoDB URL is now using a
secret in map-api.yaml instead of a plain environment variable.

This commit also adds an alternative to compose. This commit adds a new
yaml file for a MongoDB container in Kubernetes for testing/development.

This commit also moves the yaml files in the manifests folder.

This commit also modifies Travis CI configuration. It's now using the
CI scripts in IBM/pattern-ci

Fixes #9, Fixes #10

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>